Fix the mouse position on selection
We used to have a title bar, but that changed and we now either only
have a sidebar or the sidebar becomes the title bar depending on the
window size. This means that the position of the messages component is
not static any more so mouse events became wrong in certain cases (when
there is no title bar). This fixes it by sending the messages component
its absolute position on the screen when needed.

Signed-off-by: Djordje Lukic <djordje.lukic@docker.com>