Add a "Meta" property to the tool results
This makes it possible for the tools to send additional data that is not
meant to be sent to the LLM. We use this in the case of the todo toolset
to always send the current state of all of the todos. With this meta we
no longer have to parse the arguments on the TUI side.

Signed-off-by: Djordje Lukic <djordje.lukic@docker.com>